vtep: Remove 'logical_switches' column from Global table.
Links from both Physical_Port and Global to Logical_Switch raised the
possibility of a logical inconsistency in the database: every
Logical_Switch should have a link from Global, but the database schema
could not enforce this.  So, this commit removes the 'logical_switches'
column that contained this link.

On the other hand, it is reasonable to have a Logical_Switch not bound to
any Physical_Port, so this commit also adds Logical_Switch to the root set.

Signed-off-by: Ben Pfaff <blp@nicira.com>