vtep: Remove SSL table and column from schema.
Ken Duda pointed out that this will be configured out of band.

It might make sense to remove the Manager table and column for the same
reason, but that is less clear: NVP may want to add and update a set of
extra active-backup managers, with only a single one of those initially
set up through out-of-band mechanisms.  (This is what we do for Open
Open vSwitch.)

Signed-off-by: Ben Pfaff <blp@nicira.com>